backend/requirements.txt
# fastapi
# requests
# uvicorn
# pymongo
# numpy
# torch --index-url https://download.pytorch.org/whl/cpu
# python-dotenv
ipykernel
# python-multipart
farm-haystack
